// client/src/app/view/SF/SF00205/model/SF00205_Filter.model.ts
/**
 * Created by manhnv on 6/14/2017.
 */

export class SF00205Filter {
    periodType: number = 1;
    fromDate: Date;
    toDate: Date;

    selectedDepartmentId: number;
    selectedPicId: number;

    dealCode: string;
    dealName: string;

    customerCode: string;
    customerName: string;
}